libs/jervis_contracts/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jervis-contracts"
version = "0.1.0"
description = "Generated Protobuf + gRPC stubs for inter-service contracts in the Jervis monorepo."
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.66",
    "grpcio-tools>=1.66",
    "grpcio-reflection>=1.66",
    "protobuf>=5.34,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["jervis_contracts*", "jervis*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"jervis.common" = ["*.pyi"]
